create the parent class for Agitha and Jovani
(and now we see why the last patch was so so broken up)
