+++ hints/data/parse_log.py
@@ -3,6 +3,7 @@
 # take the data and handle it elsewhere.
 from hints.control.program import Program
 from hints.tabs.shopping.agitha_tab import AgithaTab
+from hints.tabs.shopping.jovani_tab import JovaniTab
 from json import load
 from pathlib import Path
 from re import findall, sub
@@ -80,4 +81,4 @@ class ParseLog:
                 # Special handling for Jovani
                 elif sign == 'Jovani_House_Sign':
                     # Go to his parsing
-                    print(hint_text)
+                    JovaniTab(self.program, hint_text)
